% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{Anscombe}
\alias{Anscombe}
\title{Anscombe quartet}
\format{A data frame with three variables:
\itemize{
\item{\strong{quartet}} {The Anscombe quartet (1th, 2th, 3th, or 4th)}
\item{\strong{x}} {numeric x variable (integers)}
\item{\strong{y}} {numeric y variable (continous)}
}}
\source{
Data gathered from \code{dataset::anscombe}
}
\usage{
Anscombe
}
\description{
Four pair-wise sets of data (y ~ x) that have the same basic statsitical properties
but greatly differ in graphical form and shape.
}
\references{
F. J. Anscombe, Graphs in Statistical Analysis. Am. Stat. 27, 17 (1973).
}
\keyword{datasets}
