% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{malaria}
\alias{malaria}
\title{Malaria deaths worldwide.}
\format{A data frame with five variables:
\itemize{
\item{\strong{country}} {The name of the country}
\item{\strong{code}} {The code of the country}
\item{\strong{year}} {The year of deaths}
\item{\strong{age_group}} {The age category}
\item{\strong{deaths}} {Number of malaria deaths per 100,000 individuals}
}}
\source{
Data downloaded from \strong{TidyTuesday Project} Github repository:
\url{https://github.com/rfordatascience/tidytuesday}
}
\usage{
malaria
}
\description{
Malaria deaths by age across the world and time.
}
\details{
For further details, see \url{https://ourworldindata.org/malaria}
}
\references{
Max Roser and Hannah Ritchie (2019) - "Malaria". Published online at OurWorldInData.org.
Retrieved from: \url{https://ourworldindata.org/malaria}.
}
\keyword{datasets}
