% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{package}
\name{teads-package}
\alias{teads-package}
\title{Worldwide datasets for teaching data science.}
\description{
An implementation of sensitivity analysis for phylogenetic comparative
An umbrella of social and environmental datasets worldwide for
teaching data science, R programming, basic statistics and data visualization in R.
The package is an umbrella of usefull datasets to be used for teaching data analysis and
statistics.
}
\details{
This package is build on the work of many that gathered this incredible datasets.
Always check details and reference to credit authors involved.
}
\section{Bug reports}{


\url{https://github.com/paternogbc/teads/issues}.
}

\section{How to contribute?}{


You can find instructions on how to contribute to \bold{teads} at
this link: \url{https://github.com/paternogbc/teads}
}

\author{
Gustavo Brant Paterno
}
