% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_data.R
\docType{data}
\name{warming}
\alias{warming}
\title{Global warming}
\format{A data frame with three variables:
\itemize{
\item{\strong{Source}} {The source of data: \strong{(1)} GISS Surface Temperature
(GISTEMP) analysis and \strong{(2)} the global component of Climate at a Glance (GCAG).}
\item{\strong{year}} {The year of temperature measures.}
\item{\strong{month}} {The month of temperature measures.}
\item{\strong{day}} {The day of temperature measures.}
\item{\strong{temp}} {Combined Land-Surface Air and Sea-Surface Water Temperature Anomalies
(deviations from a corresponding period, see details)}
}}
\source{
Data downloaded from \url{https://datahub.io/core/global-temp#readme}
}
\usage{
warming
}
\description{
Global Annual Mean Surface Air Temperature Change
}
\details{
For \strong{GISTEMP}: Temperature Anomalies are deviations from 1951-1980
means (see \url{https://data.giss.nasa.gov/gistemp})

For \strong{GCAG}: Temperature Anomalies are deviations with respect to the
20th century average (see \url{http://www.ncdc.noaa.gov/cag/data-info/global})
}
\references{
GISTEMP: NASA Goddard Institute for Space Studies (GISS) Surface
Temperature Analysis, Global Land-Ocean Temperature Index.

NOAA National Climatic Data Center (NCDC), global component of Climate
at a Glance (GCAG).
}
\keyword{datasets}
